node > node data db

+++ Libs/Repositories.Dapper/NodesStatesRepositoryDapper.cs
@@ -24,7 +24,7 @@ namespace MyNetSensors.Repositories.Dapper
         private int writeInterval = 5000;
         private Timer updateDbTimer = new Timer();
         private List<NodeState> cachedStates = new List<NodeState>();
-        private Dictionary<string, int> maxStates = new Dictionary<string, int>(); //nodeId,maxStates
+        private Dictionary<string, int?> maxStates = new Dictionary<string, int?>(); //nodeId,maxStates
 
         public event LogEventHandler OnLogInfo;
         public event LogEventHandler OnLogError;
@@ -127,23 +127,31 @@ namespace MyNetSensors.Repositories.Dapper
                     List<NodeState> statesForNode = states.Where(x => x.NodeId == nodeId).ToList();
                     states.RemoveAll(x => x.NodeId == nodeId);
 
-                    int dbCount = db.Query<int>("SELECT COUNT(*) FROM [NodesStates] WHERE NodeId=@nodeId", new { nodeId }).Single();
-
-                    int allCount = dbCount + statesForNode.Count;
-                    int max = maxStates[nodeId];
-                    int more = allCount - max;
-                    if (more > 0)
+                    //remove extra data
+                    if (maxStates[nodeId] != null)
                     {
-                        int removeFromDb = allCount - max;
-                        if (removeFromDb > 0)
-                            db.Query($"DELETE FROM [NodesStates] WHERE Id IN (SELECT TOP {more} Id FROM [NodesStates] WHERE NodeId=@nodeId ORDER BY DateTime ASC);", new { nodeId });
+                        int dbCount =
+                            db.Query<int>("SELECT COUNT(*) FROM [NodesStates] WHERE NodeId=@nodeId", new {nodeId})
+                                .Single();
 
-                        int removeFromCached = statesForNode.Count - max;
-                        if (removeFromCached > 0)
-                            statesForNode.RemoveRange(0, removeFromCached);
+                        int allCount = dbCount + statesForNode.Count;
+                        int max = maxStates[nodeId].Value;
+                        int more = allCount - max;
+                        if (more > 0)
+                        {
+                            int removeFromDb = allCount - max;
+                            if (removeFromDb > 0)
+                                db.Query(
+                                    $"DELETE FROM [NodesStates] WHERE Id IN (SELECT TOP {more} Id FROM [NodesStates] WHERE NodeId=@nodeId ORDER BY DateTime ASC);",
+                                    new {nodeId});
 
-                        statesToWrite.AddRange(statesForNode);
+                            int removeFromCached = statesForNode.Count - max;
+                            if (removeFromCached > 0)
+                                statesForNode.RemoveRange(0, removeFromCached);
+
+                        }
                     }
+                    statesToWrite.AddRange(statesForNode);
                 }
 
                 var sqlQuery = "INSERT INTO [NodesStates] (NodeId, State, DateTime) "
@@ -156,7 +164,7 @@ namespace MyNetSensors.Repositories.Dapper
             return inserts;
         }
 
-        public void AddState(NodeState state, int maxStatesCount)
+        public void AddState(NodeState state, int? maxStatesCount)
         {
             if (writeInterval != 0)
             {
@@ -174,13 +182,20 @@ namespace MyNetSensors.Repositories.Dapper
                                + "SELECT CAST(SCOPE_IDENTITY() as int)";
                 db.Query(sqlQuery, state);
 
+                //remove extra data
+                if (maxStatesCount != null)
+                {
+                    int count =
+                        db.Query<int>("SELECT COUNT(*) FROM [NodesStates] WHERE NodeId=@nodeId", new {state.NodeId})
+                            .Single();
 
-                int count = db.Query<int>("SELECT COUNT(*) FROM [NodesStates] WHERE NodeId=@nodeId", new { state.NodeId }).Single();
+                    int more = count - maxStatesCount.Value + 1;
 
-                int more = count - maxStatesCount + 1;
-
-                if (more > 0)
-                    db.Query($"DELETE FROM [NodesStates] WHERE Id IN (SELECT TOP {more} Id FROM [NodesStates] WHERE NodeId=@nodeId ORDER BY DateTime ASC);", new { state.NodeId });
+                    if (more > 0)
+                        db.Query(
+                            $"DELETE FROM [NodesStates] WHERE Id IN (SELECT TOP {more} Id FROM [NodesStates] WHERE NodeId=@nodeId ORDER BY DateTime ASC);",
+                            new {state.NodeId});
+                }
             }
         }
